valueをBigDecimalに統一

+++ vendingmachine/src/product/Product.java
@@ -1,17 +1,19 @@
 package vendingmachine.src.product;
 
+import java.math.BigDecimal;
+
 public enum Product {
 	COLA(100),
 	CANDY(65),
 	CHIPS(50);
 
-	private int value;
+	private BigDecimal value;
 
 	Product(int value) {
-		this.value = value;
+		this.value = BigDecimal.valueOf(value);
 	}
 
-	public int value() {
+	public BigDecimal value() {
 		return value;
 	};
 }
